# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: FedRAMP Gap Analysis Agent API
  version: 1.0.0
  description: |
    AI-powered FedRAMP High baseline compliance gap analysis agent that analyzes policy documents, 
    design documentation, and source code to identify compliance gaps and provide remediation recommendations.

    ## Features
    - Comprehensive gap analysis across policy, design, and code
    - Support for 325+ FedRAMP High baseline controls (NIST 800-53 Rev 5)
    - Risk assessment with severity classification
    - Automated remediation recommendations with code examples
    - Multiple report formats (JSON, PDF, HTML, Excel)

    ## Authentication
    All API endpoints require authentication using Bearer token or API key.

  contact:
    name: API Support
    email: support@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.fedramp-agent.example.com/api/v1
    description: Production server
  - url: https://staging-api.fedramp-agent.example.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development server
...
paths:
...
  /controls:
    get:
      tags:
        - Controls
      summary: List FedRAMP controls
      description: Get list of all FedRAMP High baseline controls
      operationId: listControls
      parameters:
        - name: family
          in: query
          schema:
            type: string
          description: Filter by control family (AC, AU, IA, etc.)
        - name: baseline
          in: query
          schema:
            type: string
            enum: [Low, Moderate, High]
          description: Filter by baseline
        - name: search
          in: query
          schema:
            type: string
          description: Search query
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Controls retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlListResponse"
      security:
        - BearerAuth: []
  /controls/{controlId}:
    get:
      tags:
        - Controls
      summary: Get control details
      description: Get detailed information about a specific control
      operationId: getControl
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
          description: Control identifier (e.g., AC-2, IA-2(1))
      responses:
        "200":
          description: Control details retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
        "404":
          $ref: "#/components/responses/NotFound"
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
...
    Control:
      type: object
      properties:
        control_id:
          type: string
        control_name:
          type: string
        control_family:
          type: string
        baseline:
          type: string
        description:
          type: string
        implementation_guidance:
          type: string
        patterns:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        verification_methods:
          type: array
          items:
            type: string
    ControlListResponse:
      type: object
      properties:
        total:
          type: integer
        controls:
          type: array
          items:
            $ref: "#/components/schemas/Control"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...